+++ src/main/kotlin/app/revanced/patches/reddit/customclients/slide/api/SpoofClientPatch.kt
@@ -9,7 +9,7 @@ import app.revanced.patches.reddit.customclients.AbstractSpoofClientPatch
 import app.revanced.patches.reddit.customclients.boostforreddit.api.fingerprints.GetClientIdFingerprint
 
 @Patch(
-    name = "Slide client spoof",
+    name = "Spoof client",
     description = "Spoofs the client in order to allow logging in. " +
             "The OAuth application type has to be \"Installed app\" " +
             "and the redirect URI has to be set to \"http://www.ccrama.me\".",