+++ src/main/kotlin/app/revanced/patches/strava/subscription/UnlockSubscriptionPatch.kt
@@ -10,7 +10,7 @@ import app.revanced.patches.strava.subscription.fingerprints.GetSubscribedFinger
 
 @Patch(
     name = "Unlock subscription features",
-    description = "Unlocks \"Matched Runs\" and \"Segment Efforts\".",
+    description = "Unlocks \"Routes\", \"Matched Runs\" and \"Segment Efforts\".",
     compatiblePackages = [CompatiblePackage("com.strava", ["320.12"])]
 )
 @Suppress("unused")